requirements_module_n.txt
# Module N - File Ingest & Extract System
# Additional dependencies for Module N service

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Excel Processing
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# PDF Processing
PyMuPDF==1.23.8
camelot-py==0.11.0
tabula-py==2.9.0
pdf2image==1.16.3

# OCR
pytesseract==0.3.10

# File Type Detection
python-magic==0.4.27
python-magic-bin==0.4.14  # Windows binary

# HTTP Client (for webhooks)
requests==2.31.0
httpx==0.25.2

# Database (already in main requirements.txt, listed for reference)
# SQLAlchemy==2.0.23
# ezdxf==1.1.0
# Pillow==10.1.0
# python-dotenv==1.0.0

# Development & Testing
pytest-asyncio==0.21.1

